requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
celery-progress==0.0.3
certifi==2019.3.9
chardet==3.0.4
croniter==0.3.30
cycler==0.10.0
decorator==4.4.0
Django==2.1
django-admin-colors==0.2.0
django-betterforms==1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==1.4.2
django-bootstrap4==0.0.8
django-celery-beat==1.5.0
django-crontab==0.7.1
django-export==1.11.0
django-form-utils==1.0.3
django-leaflet==0.24.0
django-modeladmin-reorder==0.3.1
django-reset-migrations==0.3.1
django-tempus-dominus==5.1.2.5
django-timezone-field==3.0
flower==0.9.3
future==0.17.1
idna==2.8
kiwisolver==1.1.0
kombu==4.6.1
lxml==4.3.4
matplotlib==3.1.0
mysql==0.0.2
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
numpy==1.16.4
obspy==1.1.1
protobuf==3.8.0
pyparsing==2.4.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.4
tzlocal==2.0.0
urllib3==1.25.3
vine==1.3.0
